Enable JSON subscription imports (#878)
* Rename parser variable

* Add support for parsing json files

The files are now parsed as json if xml is not explicitly specified
in the file suffix. The old xml parser is kept more or less intact.

* Remove XML from GUI strings

Mentions of "xml" have been removed from all strings related to
subscription imports. I hope that i didn't butcher any translations.

* Fix error in if clause

* Close file reader after reading

* Improve readability of null checking

* Turn JSON primitives into Strings properly

* Revert "Remove XML from GUI strings"

This reverts commit a9165a8052281194f2703656d59a6348ae3acffa.
Following advice from maintainer, the changes to language files
should be reverted and left to translators.

* Remove mentions of XML from base language file

This change was previously reverted along with changes to the
translation files. This change is still desirable, however.

* Add json to comments mentioning xml files

* Fix Swedish translation

As a Swedish native speaker, I felt that I was at least qualified
to edit this translation. If it should be done by somebody else
and/or in another way, this commit may be omitted from the merge.

* Revert "Remove mentions of XML from base language file"

This reverts commit 73287adcc5ab79399e9b751fe2e5d965eb73cad8.
This change was not desirable, after all.

* Change string name and remove XML from GUI strings

Changed string name to be more general and not include file type.
The reference is also changed in the Swedish language file, as the
translations there have also been changed in this PR. References in
other translation files need to be changed by their respective
translators. The word XML has been removed from GUI strings where
applicable.

* Fix forgotten reference in code